requirements.txt
zai-sdk
python-dotenv
requests
flask
flask-socketio
openai
pillow
matplotlib
plotly
pandas
numpy
beautifulsoup4
tweepy
python-linkedin-v2
instagrapi
websockets
aiofiles
sqlite3